requirements.txt
flask==3.0.0
flask-socketio==5.3.6
ollama==0.1.6
python-dotenv==1.0.0
python-engineio==4.9.0
python-socketio==5.10.0